Cleanup class names
